EDGECLOUD-1829 MC giving quota exceeded error on CreateUser (#621)
Preparation for switching MC emails from Gmail to Sendgrid.  MC
currently assumes that the SMTP server login username is the same as the
email address, which is not true for Sendgrid.

The additional "user" parameter has been added to the "noreplyemail"
account in main, qa, stage, and dev vaults, and set to be the same as
"email" for Gmail.