EDGECLOUD-3828 Continuous Scanning for CVE based updates for base image packages (#1124)
This change has a few components:

- The Ubuntu APT cache server (apt.mobiledgex.net) which uses "aptly" to create stable snapshots of upstream Ubuntu packages.
- Changes to the base image to use these snapshots for package installation.
- Chef changes to upgrade a base image from one snapshot to another. (TBD)
- An offline process to scan the list of packages for each base image periodically for vulnerabilities.

The APT cache is ready and currently hosts the snapshot: https://apt.mobiledgex.net/cirrus/2020-10-22. The changes in this commit switches the 4.1.1 base image to use this snapshot.

The CIS scan Jenkins job which we run for each base image now also stashes the package list for the base image in Artifactory. An offline process running in roma.mobiledgex.net does a twice daily scan of these packages using Ubuntu's package repository and alerts on packages which need updates.

The chef task to update VMs to newer snapshots will be added in a subsequent commit, once we have moved to the 4.1.1 base image.

An unrelated change in the mobiledgex package dependencies files relaxes dependencies for non-critical packages and specifies lower-bound (">=") dependencies instead of hard single-version dependencies. This makes it easier to upgrade the mobiledgex package.