Optimized Metrics API (#1482)
Allow multiple app instances be specified for app metrics queries. Changed the format for the returned data in order to make it more compact by grouping the data on the per-appInst basis. In addition added downsampling to the returned data in order for it to be more friendly in the UI.