EDGECLOUD-5529 EDGECLOUD-5534 Check for startage/endage. Change Port from value to a tag (#1717)
EDGECLOUD-5534
Change port from value to a tag, if it is not part of the key it would overwrite the previous value and only one port would be stored per appInstance. This change will require reset of influxDB connections measurement. Also might require some changes in the UI as well.

EDGECLOUD-5529
Check for startage and endage to be not set before skipping Resolve call.
Added unit-test for it