Cluster and cloudlet grouped metrics api (#1716)
This change adds a grouped metrics api for cluster and cloudlet metrics. This is similar to grouped app metrics doesn't effect existing apis right now, but once UI transitions to the new api the old api will be removed. The change groups the metrics and as a result the output is a lot more compact and is much easier to process for the UI.

Refer to the examples page for the new api format: https://mobiledgex.atlassian.net/wiki/spaces/SWDEV/pages/2266398741/Grouped+metrics+API

As part of the change I refactored the metrics code to use an interface to make most of the query building code.

Also added unit-tests for grouped metrics for cluster and cloudlet.