Switch to component-specific docker images (#1499)
All components get their own individual images which include binaries picked from the all-in-one edge-cloud docker image. The base for most of the images is the GoogleContainerTools distroless base-debian10 image which is a minimal image with nothing apart from libc, timezone data, and CA certificates. Our current images are built with CGO as a bunch of the binaries use plugins, and we need a base image which has libc at least.

CRM and shepherd share a container which is built from our current edge-cloud-base-image as they have a bunch of dependencies on external tools. The controller also needs the edge-cloud-base-image as it uses IaaS CLIs like openstack, gcloud, az.