Re-build shared rootLb state  (#342)
Added code to re-build shared rootLb state, if the previous one was deleted. When the crm comes up it does one of the two things - updates rootLb VM, or creates it. If shared rootLb was created, we check if there are any clusters on this cloudlet, that use the shared rootLb, and if there are several things are done(CheckRebuildRootLb):
* Update rootLb vm with the cluster network
* update rootLb vm with kubeconfig for the cluster
* Walk the appInsts to rebuild manifest states for all the created appInsts on the shared cloudlet.
* Re-create all the envoy/nginx proxies for appInsts
* Refactored some code for better reuse.