Crm debug to get logs (#443)
In the restricted cloudlet environment it's difficult to get the logs from crm. This change adds a new command
"mcctl debug rundebug cmd=crmlogs args="<grep args>" cloudlet=<cloudletname>"
It runs docker logs crmserver on the platform vm. Command also takes a string as an arg, which will be used to grep the logfile for(docker logs crmserver 2>&1 | grep "<args>").
NOTE: cloudlet name is mandatory, due to the potential of the command to return a very large amounts of data.
In addition I added more logs for when the k8s cluster create fails.